Update epic.md
Add Yaping Dan into VE320, who promised to give A- or more to students who achieved 70 in overall in semester, while eventually a student who got more than 80 with 3 Q3 in exams and almost full grade in quiz and homework also got A- (usually 10 points stands for 2 letter grade variance, i.e A- to A+). No judgement on not giving good grade, but it's kinda "PUA" for not keeping the promise to some extents.